Fixes redundancy
